# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/quick-test/sessions": {
      "post": {
        "tags": [
          "quick-test"
        ],
        "summary": "Create Quick Test Session",
        "operationId": "create_quick_test_session_quick_test_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuickTestCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuickTestStartResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/quick-test/sessions/{session_id}/answer": {
      "post": {
        "tags": [
          "quick-test"
        ],
        "summary": "Answer Quick Test",
        "operationId": "answer_quick_test_quick_test_sessions__session_id__answer_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuickTestAnswerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuickTestAnswerResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/quick-test/sessions/{session_id}": {
      "get": {
        "tags": [
          "quick-test"
        ],
        "summary": "Get Quick Test Session",
        "operationId": "get_quick_test_session_quick_test_sessions__session_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuickTestSessionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/quick-test/sessions/{session_id}/finish": {
      "post": {
        "tags": [
          "quick-test"
        ],
        "summary": "Finish Quick Test Session",
        "operationId": "finish_quick_test_session_quick_test_sessions__session_id__finish_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuickTestFinishResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QuickTestAnswerRequest": {
        "properties": {
          "answer": {
            "type": "string",
            "title": "Answer"
          }
        },
        "type": "object",
        "required": [
          "answer"
        ],
        "title": "QuickTestAnswerRequest"
      },
      "QuickTestAnswerResponse": {
        "properties": {
          "feedback": {
            "type": "string",
            "title": "Feedback"
          },
          "score": {
            "type": "integer",
            "title": "Score"
          },
          "next_question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Question"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/SourceCitation"
            },
            "type": "array",
            "title": "Citations",
            "default": []
          },
          "is_finished": {
            "type": "boolean",
            "title": "Is Finished"
          }
        },
        "type": "object",
        "required": [
          "feedback",
          "score",
          "is_finished"
        ],
        "title": "QuickTestAnswerResponse"
      },
      "QuickTestCreateRequest": {
        "properties": {
          "source_type": {
            "type": "string",
            "enum": [
              "note",
              "knowledge",
              "mixed"
            ],
            "title": "Source Type"
          },
          "source_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Source Ids"
          },
          "question_count": {
            "type": "integer",
            "title": "Question Count",
            "default": 5
          },
          "difficulty": {
            "type": "string",
            "enum": [
              "easy",
              "normal",
              "hard"
            ],
            "title": "Difficulty",
            "default": "normal"
          },
          "focus": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Focus"
          }
        },
        "type": "object",
        "required": [
          "source_type",
          "source_ids"
        ],
        "title": "QuickTestCreateRequest"
      },
      "QuickTestFinishResponse": {
        "properties": {
          "final_summary": {
            "type": "string",
            "title": "Final Summary"
          },
          "weak_points": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Weak Points",
            "default": []
          },
          "recommended_notes": {
            "items": {
              "$ref": "#/components/schemas/SourceCitation"
            },
            "type": "array",
            "title": "Recommended Notes",
            "default": []
          },
          "recommended_documents": {
            "items": {
              "$ref": "#/components/schemas/SourceCitation"
            },
            "type": "array",
            "title": "Recommended Documents",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "final_summary"
        ],
        "title": "QuickTestFinishResponse"
      },
      "QuickTestSessionResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type"
          },
          "source_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Source Ids"
          },
          "question_count": {
            "type": "integer",
            "title": "Question Count"
          },
          "difficulty": {
            "type": "string",
            "title": "Difficulty"
          },
          "focus": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Focus"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "current_turn": {
            "type": "integer",
            "title": "Current Turn"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "weak_points": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Weak Points",
            "default": []
          },
          "recommended_refs": {
            "items": {
              "$ref": "#/components/schemas/SourceCitation"
            },
            "type": "array",
            "title": "Recommended Refs",
            "default": []
          },
          "turns": {
            "items": {
              "$ref": "#/components/schemas/QuickTestTurnResponse"
            },
            "type": "array",
            "title": "Turns",
            "default": []
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "source_type",
          "source_ids",
          "question_count",
          "difficulty",
          "status",
          "current_turn"
        ],
        "title": "QuickTestSessionResponse"
      },
      "QuickTestStartResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "first_question": {
            "type": "string",
            "title": "First Question"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/SourceCitation"
            },
            "type": "array",
            "title": "Citations",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "first_question"
        ],
        "title": "QuickTestStartResponse"
      },
      "QuickTestTurnResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "turn_index": {
            "type": "integer",
            "title": "Turn Index"
          },
          "question": {
            "type": "string",
            "title": "Question"
          },
          "answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer"
          },
          "feedback": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback"
          },
          "score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/SourceCitation"
            },
            "type": "array",
            "title": "Citations",
            "default": []
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "turn_index",
          "question"
        ],
        "title": "QuickTestTurnResponse"
      },
...
      "SourceCitation": {
        "properties": {
          "source_type": {
            "type": "string",
            "title": "Source Type"
          },
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "chunk_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chunk Id"
          },
          "quote": {
            "type": "string",
            "title": "Quote"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "source_type",
          "source_id",
          "title",
          "quote"
        ],
        "title": "SourceCitation"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}